dar resultado en tipo flotante
se cambio a tipo flotante para manterner un estandar en las operaciones aritmeticas

+++ Operaciones/aritmetica.py
@@ -1,19 +1,19 @@
 def suma(num1,num2):
-    sumar = num1 + num2
+    sumar = float(num1 + num2)
     print('• Sumando:   ',str(num1),' + ',str(num2), ' = ', sumar)
     return sumar
 
 def resta(num1,num2):
-    restar = num1 - num2
+    restar = float(num1 - num2)
     print('• Restando:  ',str(num1),' - ',str(num2), ' = ', restar)
     return restar
 
 def multiplicacion(num1,num2):
-    multiplicar = num1 * num2
+    multiplicar = float(num1 * num2)
     print('• Multiplicando:  ',str(num1),' * ',str(num2), ' = ', multiplicar)
     return multiplicar
 
 def division(num1,num2):
-    dividir = num1 / num2
+    dividir = float(num1 / num2)
     print('• Dividiendo:  ',str(num1),' / ',str(num2), ' = ', dividir)
     return dividir
